refactor: extract tests for core CLI commands to test_core.py
