+++ tests/cli/test_cli.py
@@ -15,13 +15,8 @@
 import io
 import math
 import os
-import random
-import string
 import subprocess
-import sys
-import textwrap
 from pathlib import Path
-from unittest import TestCase
 
 import pytest
 import yaml
@@ -30,41 +25,6 @@ from click.testing import CliRunner
 from peekingduck import __version__
 from peekingduck.cli import cli
 from peekingduck.declarative_loader import PEEKINGDUCK_NODE_TYPES
-from tests.conftest import assert_msg_in_logs
-
-UNIQUE_SUFFIX = "".join(random.choice(string.ascii_lowercase) for _ in range(8))
-CUSTOM_FOLDER_NAME = f"custom_nodes_{UNIQUE_SUFFIX}"
-
-PKD_NODE_TYPE = "pkd_node_type"
-PKD_NODE_NAME = "pkd_node_name"
-PKD_NODE_NAME_2 = "pkd_node_name2"
-PKD_NODE = f"{CUSTOM_FOLDER_NAME}.{PKD_NODE_TYPE}.{PKD_NODE_NAME}"
-PKD_NODE_2 = f"{CUSTOM_FOLDER_NAME}.{PKD_NODE_TYPE}.{PKD_NODE_NAME_2}"
-NODES = {"nodes": [PKD_NODE, PKD_NODE_2]}
-
-PROJECT_DIR = Path("tmp_dir")
-MODULE_DIR = PROJECT_DIR / "src"
-
-DEFAULT_NODE_DIR = MODULE_DIR / "custom_nodes"
-DEFAULT_NODE_CONFIG_DIR = DEFAULT_NODE_DIR / "configs"
-CUSTOM_NODE_DIR = MODULE_DIR / CUSTOM_FOLDER_NAME
-CUSTOM_NODE_CONFIG_DIR = CUSTOM_NODE_DIR / "configs"
-CUSTOM_PKD_NODE_DIR = MODULE_DIR / CUSTOM_FOLDER_NAME / PKD_NODE_TYPE
-CUSTOM_PKD_NODE_CONFIG_DIR = CUSTOM_NODE_CONFIG_DIR / PKD_NODE_TYPE
-PIPELINE_PATH = Path("pipeline_config.yml")
-CUSTOM_PIPELINE_PATH = Path("custom_dir") / "pipeline_config.yml"
-YML = dict(
-    nodes=[
-        {
-            "input.visual": {
-                "source": "https://storage.googleapis.com/peekingduck/videos/wave.mp4"
-            }
-        },
-        "model.posenet",
-        "draw.poses",
-        "output.screen",
-    ]
-)
 
 PKD_DIR = Path(__file__).resolve().parents[2] / "peekingduck"
 PKD_CONFIG_DIR = PKD_DIR / "configs"
@@ -98,82 +58,6 @@ def available_nodes_msg(type_name=None):
     return content
 
 
-def create_node_config(config_dir, node_name, config_text):
-    with open(config_dir / f"{node_name}.yml", "w") as outfile:
-        yaml.dump(config_text, outfile)
-
-
-def create_node_python(node_dir, node_name, return_statement):
-    with open(node_dir / f"{node_name}.py", "w") as outfile:
-        content = textwrap.dedent(
-            f"""\
-            from peekingduck.pipeline.nodes.abstract_node import AbstractNode
-
-            class Node(AbstractNode):
-                def __init__(self, config, **kwargs):
-                    super().__init__(config, node_path=__name__, **kwargs)
-
-                def run(self, inputs):
-                    return {{ {return_statement} }}
-            """
-        )
-        outfile.write(content)
-
-
-def create_pipeline_yaml(nodes, custom_config_path):
-    with open(
-        CUSTOM_PIPELINE_PATH if custom_config_path else PIPELINE_PATH, "w"
-    ) as outfile:
-        yaml.dump(nodes, outfile, default_flow_style=False)
-
-
-def get_custom_node_subpaths(node_subdir, node_type, node_name):
-    return (
-        str(Path(node_subdir) / "configs" / node_type / f"{node_name}.yml"),
-        str(Path(node_subdir) / node_type / f"{node_name}.py"),
-    )
-
-
-def init_msg(node_name):
-    return f"Initializing {node_name} node..."
-
-
-def setup_custom_node(node_subdir, node_type, node_name):
-    cwd = Path.cwd()
-    config_dir = cwd / node_subdir / "configs" / node_type
-    script_dir = cwd / node_subdir / node_type
-    config_dir.mkdir(parents=True, exist_ok=True)
-    script_dir.mkdir(parents=True, exist_ok=True)
-
-    (config_dir / f"{node_name}.yml").touch()
-    (script_dir / f"{node_name}.py").touch()
-
-
-def setup(custom_config_path=False):
-    sys.path.append(str(MODULE_DIR))
-
-    relative_node_dir = CUSTOM_PKD_NODE_DIR.relative_to(CUSTOM_PKD_NODE_DIR.parts[0])
-    relative_config_dir = CUSTOM_PKD_NODE_CONFIG_DIR.relative_to(
-        CUSTOM_PKD_NODE_CONFIG_DIR.parts[0]
-    )
-    relative_node_dir.mkdir(parents=True)
-    relative_config_dir.mkdir(parents=True)
-    if custom_config_path:
-        CUSTOM_PIPELINE_PATH.parent.mkdir()
-    node_config_1 = {
-        "input": ["none"],
-        "output": ["test_output_1"],
-        "resize": {"do_resizing": False},
-    }
-    node_config_2 = {"input": ["test_output_1"], "output": ["pipeline_end"]}
-
-    create_pipeline_yaml(NODES, custom_config_path)
-    create_node_python(relative_node_dir, PKD_NODE_NAME, "'test_output_1': None")
-    create_node_python(relative_node_dir, PKD_NODE_NAME_2, "'pipeline_end': True")
-    create_node_config(relative_config_dir, PKD_NODE_NAME, node_config_1)
-    create_node_config(relative_config_dir, PKD_NODE_NAME_2, node_config_2)
-
-
 @pytest.fixture
 def cwd():
     return Path.cwd()
@@ -193,83 +77,6 @@ class TestCli:
         # not testing full message as .invoke() sets program name to cli
         # instead of peekingduck
         assert f"version {__version__}" in result.output
-
-    def test_init_default(self, parent_dir, cwd):
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            result = CliRunner().invoke(cli, ["init"])
-
-            assert result.exit_code == 0
-            assert_msg_in_logs(
-                f"Creating custom nodes folder in {cwd / 'src' / 'custom_nodes'}",
-                captured.records,
-            )
-            assert (parent_dir / DEFAULT_NODE_DIR).exists()
-            assert (parent_dir / DEFAULT_NODE_CONFIG_DIR).exists()
-            assert (cwd / PIPELINE_PATH).exists()
-            with open(cwd / PIPELINE_PATH) as infile:
-                config_file = yaml.safe_load(infile)
-                TestCase().assertDictEqual(YML, config_file)
-
-    def test_init_custom(self, parent_dir, cwd):
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            result = CliRunner().invoke(
-                cli, ["init", "--custom_folder_name", CUSTOM_FOLDER_NAME]
-            )
-
-            assert result.exit_code == 0
-            assert_msg_in_logs(
-                f"Creating custom nodes folder in {cwd / 'src' / CUSTOM_FOLDER_NAME}",
-                captured.records,
-            )
-            assert (parent_dir / CUSTOM_NODE_DIR).exists()
-            assert (parent_dir / CUSTOM_NODE_CONFIG_DIR).exists()
-            assert (cwd / PIPELINE_PATH).exists()
-            with open(cwd / PIPELINE_PATH) as infile:
-                TestCase().assertDictEqual(YML, yaml.safe_load(infile))
-
-    def test_run_default(self):
-        setup()
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            result = CliRunner().invoke(cli, ["run"])
-
-            assert_msg_in_logs("Successfully loaded pipeline file.", captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE), captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE_2), captured.records)
-
-            assert result.exit_code == 0
-
-    def test_run_custom_path(self):
-        setup(True)
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            result = CliRunner().invoke(
-                cli, ["run", "--config_path", CUSTOM_PIPELINE_PATH]
-            )
-
-            assert_msg_in_logs("Successfully loaded pipeline file.", captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE), captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE_2), captured.records)
-            assert result.exit_code == 0
-
-    def test_run_custom_config(self):
-        setup()
-        node_name = ".".join(PKD_NODE.split(".")[1:])
-        config_update_value = "'do_resizing': True"
-        config_update_cli = (
-            f"{{'{node_name}': {{'resize': {{ {config_update_value} }} }} }}"
-        )
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            result = CliRunner().invoke(
-                cli, ["run", "--node_config", config_update_cli]
-            )
-
-            assert_msg_in_logs("Successfully loaded pipeline file.", captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE), captured.records)
-            assert_msg_in_logs(
-                f"Config for node {node_name} is updated to: {config_update_value}",
-                captured.records,
-            )
-            assert_msg_in_logs(init_msg(PKD_NODE_2), captured.records)
-            assert result.exit_code == 0
 
     def test_nodes_all(self):
         result = CliRunner().invoke(cli, ["nodes"])
@@ -321,15 +128,3 @@ class TestCli:
         exit_status = proc.returncode
         assert "DEBUG" in out_str
         assert exit_status == 0
-
-    def test_num_iter(self):
-        setup()
-        with TestCase.assertLogs("peekingduck.cli.logger") as captured:
-            n = 50  # run test for 50 iterations
-            result = CliRunner().invoke(cli, ["run", "--num_iter", n])
-
-            assert_msg_in_logs("Successfully loaded pipeline file.", captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE), captured.records)
-            assert_msg_in_logs(init_msg(PKD_NODE_2), captured.records)
-            assert_msg_in_logs(f"Run pipeline for {n} iterations", captured.records)
-            assert result.exit_code == 0
